iris/requirements-dev.txt
pandas
sqlalchemy-iris==0.10.5
sentence_transformers
intersystems-irispython